Moves function from common to util
