+++ inc/t_cose/t_cose_common.h
@@ -15,8 +15,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-#include <qcbor/qcbor.h>
-
+#include "t_cose/q_useful_buf.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -670,29 +669,6 @@ struct t_cose_parameters {
 #define T_COSE_OPT_DECODE_ONLY  0x00000008
 
 /**
- * \brief Map QCBOR decode error to COSE errors.
- *
- * \param[in] qcbor_error   The QCBOR error to map.
- *
- * \return This returns one of the error codes defined by
- *         \ref t_cose_err_t.
- */
-static inline enum t_cose_err_t
-qcbor_decode_error_to_t_cose_error(QCBORError qcbor_error)
-{
-    if(qcbor_error == QCBOR_ERR_TOO_MANY_TAGS) {
-        return T_COSE_ERR_TOO_MANY_TAGS;
-    }
-    if(QCBORDecode_IsNotWellFormedError(qcbor_error)) {
-        return T_COSE_ERR_CBOR_NOT_WELL_FORMED;
-    }
-    if(qcbor_error != QCBOR_SUCCESS) {
-        return T_COSE_ERR_SIGN1_FORMAT;
-    }
-    return T_COSE_SUCCESS;
-}
-
-/**
  * \brief  Check whether an algorithm is supported.
  *
  * \param[in] cose_algorithm_id        COSE Integer algorithm ID.
@@ -708,8 +684,6 @@ qcbor_decode_error_to_t_cose_error(QCBORError qcbor_error)
  */
 bool
 t_cose_is_algorithm_supported(int32_t cose_algorithm_id);
-
-
 
 #ifdef __cplusplus
 }
