+++ src/t_cose_util.h
@@ -2,7 +2,7 @@
  *  t_cose_util.h
  *
  * Copyright 2019-2021, Laurence Lundblade
- * Copyright (c) 2020, Arm Limited. All rights reserved.
+ * Copyright (c) 2020-2022, Arm Limited. All rights reserved.
  *
  * SPDX-License-Identifier: BSD-3-Clause
  *
@@ -269,6 +269,29 @@ enum t_cose_err_t create_tbs_hash(int32_t                cose_algorithm_id,
  */
 struct q_useful_buf_c get_short_circuit_kid(void);
 #endif
+
+/**
+ * \brief Map QCBOR decode error to COSE errors.
+ *
+ * \param[in] qcbor_error   The QCBOR error to map.
+ *
+ * \return This returns one of the error codes defined by
+ *         \ref t_cose_err_t.
+ */
+static inline enum t_cose_err_t
+qcbor_decode_error_to_t_cose_error(QCBORError qcbor_error)
+{
+    if(qcbor_error == QCBOR_ERR_TOO_MANY_TAGS) {
+        return T_COSE_ERR_TOO_MANY_TAGS;
+    }
+    if(QCBORDecode_IsNotWellFormedError(qcbor_error)) {
+        return T_COSE_ERR_CBOR_NOT_WELL_FORMED;
+    }
+    if(qcbor_error != QCBOR_SUCCESS) {
+        return T_COSE_ERR_SIGN1_FORMAT;
+    }
+    return T_COSE_SUCCESS;
+}
 
 #ifdef __cplusplus
 }
